'seguridad'

+++ app/Http/Controllers/SegGrupoController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\SegGrupoRequest;
 use App\Repositories\seg_grupo\ISegGrupoRepository;
 use Exception;
+use Illuminate\Support\Facades\DB;
 
 class SegGrupoController extends Controller
 {
@@ -35,8 +36,10 @@ class SegGrupoController extends Controller
     public function save(SegGrupoRequest $request)
     {
         try {
+            DB::beginTransaction();
             $response = $this->grupo->save($request->validated());
             if ($response) :
+                DB::commit();
                 return response()->json([
                     'type' => 'success',
                     'message' => 'Registro guardado con éxito.',
@@ -46,6 +49,7 @@ class SegGrupoController extends Controller
 
             throw new Exception('No se puedo procesar el registro.', 500);
         } catch (Exception $ex) {
+            DB::rollback();
             return response()->json(['type' => 'error', 'message' => $ex->getMessage()], 500);
         }
     }
@@ -53,8 +57,14 @@ class SegGrupoController extends Controller
     public function update(SegGrupoRequest $request, $id)
     {
         try {
+            $grupo = $this->grupo->getGrupo($id);
+            if (!$grupo)
+                throw new Exception('No se encontro el registro', 404);
+
+            DB::beginTransaction();
             $response = $this->grupo->update(array_merge($request->input(), ['id' => $id]));
             if ($response) :
+                DB::commit();
                 return response()->json([
                     'type' => 'success',
                     'message' => 'Registro actualizados con éxito.'
@@ -63,6 +73,7 @@ class SegGrupoController extends Controller
 
             throw new Exception('No se puedo procesar el registro.', 500);
         } catch (Exception $ex) {
+            DB::rollback();
             return response()->json(['type' => 'error', 'message' => $ex->getMessage()], 500);
         }
     }
@@ -70,8 +81,14 @@ class SegGrupoController extends Controller
     public function delete($id)
     {
         try {
+            $grupo = $this->grupo->getGrupo($id);
+            if (!$grupo)
+                throw new Exception('No se encontro el registro', 404);
+
+            DB::beginTransaction();
             $response = $this->grupo->delete($id);
             if ($response) :
+                DB::commit();
                 return response()->json([
                     'type' => 'success',
                     'message' => 'Registro eliminado con éxito.'
@@ -80,6 +97,7 @@ class SegGrupoController extends Controller
 
             throw new Exception('No se puedo eliminar el registro.', 500);
         } catch (Exception $ex) {
+            DB::rollback();
             return response()->json(['type' => 'error', 'message' => $ex->getMessage()], 500);
         }
     }
